--- src/js/allPagination/chunkArray.js
@@ -1,9 +1,0 @@
-const chunkNewsArr = [];
-
-function chunkArray(arrayData, chunkSize) {
-  while (arrayData.length) {
-    chunkNewsArr.push(arrayData.splice(0, chunkSize));
-  }
-}
-
-export {chunkNewsArr, chunkArray}
