+++ src/js/getPopoularProduct.js
@@ -1,33 +1,68 @@
 import { getPopularNewsAPI } from './api/news-api.js';
 import { getMarkupWeather } from './markups/weather-markup.js';
 import { weatherData } from './markups/weather-markup.js';
+import { pagination } from './pagination.js';
 import { markup } from './markups/newsCard.js';
-// import {  createNewsListMarkup} from "./markups/newslistMarkup.js";
-
-
-
-import {
-    ref,
-    valuePage,
-    pagination,
-    handleButtonRight,
-    handleButtonLeft,
-    handleButton,
-  } from './allPagination/dynamicPagination.js';
-import { chunkNewsArr, chunkArray } from './allPagination/chunkArray.js';
 
 const popularNewsGallery = document.querySelector('.news-gallery');
 const notFoundPage = document.querySelector('.not-found');
+const paginationEl = document.getElementById('pagination');
+
+const valuePage = {
+  curPage: 1,
+  numLinksTwoSide: 1,
+  amountCards: 0,
+  totalPages: 0,
+};
 
 async function getPopularProduct() {
   try {
     const getNews = await getPopularNewsAPI();
     // console.log('Arr objects with mostpopular News', getNews.results);
-    newsArr = getNews.results;
+    const newsArr = getNews.results;
 
-    getAmountCards(newsArr);
-    createMarkup(newsArr);
+    let markupNews = '';
+    const markupWeather = getMarkupWeather({ data: weatherData });
+    // console.log(markupWeather);
+    // console.log({ data: weatherData });
+    const itemWeather = `<li class="weather__card">${markupWeather}</li>`;
+    // console.log(itemWeather);
 
+    if (window.innerWidth < 768) {
+      for (let i = 0; i < 5; i += 1) {
+        if (i === 0) {
+          markupNews += itemWeather;
+        } else {
+          markupNews += markup(newsArr[i]);
+        }
+        valuePage.amountCards = 5;
+        valuePage.totalPages = Math.ceil(newsArr.length / valuePage.amountCards);
+      }
+    }
+    if (window.innerWidth >= 768 && window.innerWidth < 1280) {
+      for (let i = 0; i < 8; i += 1) {
+        if (i === 1) {
+          markupNews += itemWeather;
+        } else {
+          markupNews += markup(newsArr[i]);
+        }
+        valuePage.amountCards = 8;
+        valuePage.totalPages = Math.ceil(newsArr.length / valuePage.amountCards);
+      }
+    }
+    if (window.innerWidth >= 1280) {
+      for (let i = 0; i < 9; i += 1) {
+        if (i === 2) {
+          markupNews += itemWeather;
+        } else {
+          markupNews += markup(newsArr[i]);
+        }
+        valuePage.amountCards = 9;
+        valuePage.totalPages = Math.ceil(newsArr.length / valuePage.amountCards);
+      }
+    }
+    popularNewsGallery.innerHTML = markupNews;
+    // console.log(markupNews);
     pagination(valuePage);
   } catch (error) {
     notFoundPage.classList.toggle('visually-hidden');
@@ -35,170 +70,87 @@ async function getPopularProduct() {
 }
 
 getPopularProduct();
-window.addEventListener('resize', getPopularProduct);
 
-function createMarkup(array) {
-  let markupNews = '';
-  const markupWeather = getMarkupWeather({ data: weatherData });
-  // console.log(markupWeather);
-  // console.log({ data: weatherData });
-  const itemWeather = `<li class="weather__card">${markupWeather}</li>`;
-  // console.log(itemWeather);
 
-  if (window.innerWidth < 768) {
-    for (let i = 0; i < 5; i += 1) {
-      if (i === 0) {
-        markupNews += itemWeather;
-      } else {
-        markupNews += markup(array[i]);
-      }
-    }
-  }
-  if (window.innerWidth >= 768 && window.innerWidth < 1280) {
-    for (let i = 0; i < 8; i += 1) {
-      if (i === 1) {
-        markupNews += itemWeather;
-      } else {
-        markupNews += markup(array[i]);
-      }
-    }
-  }
-  if (window.innerWidth >= 1280) {
-    for (let i = 0; i < 9; i += 1) {
-      if (i === 2) {
-        markupNews += itemWeather;
-      } else {
-        markupNews += markup(array[i]);
-      }
-    }
-  }
-  popularNewsGallery.innerHTML = markupNews;
-  // console.log(markupNews);
-}
-
-ref.paginationEl.addEventListener('click', e => {
-  const ele = e.target;
-  // console.log(ele);
-
-  if (ele.dataset.page) {
-    const pageNumber = parseInt(e.target.dataset.page, 10);
-    valuePage.curPage = pageNumber;
-    console.log(pageNumber);
-  }
-  console.log(valuePage.amountCards);
-
-  getPopularProductData()
-    .then(array => {
-      renderNewsMarkup(array, valuePage.amountCards);
-    })
-    .catch(error => console.error(error));
-  // renderNewsMarkup(newsArr, valuePage.amountCards);
-  // goToTop();
-});
-
-ref.paginationContainerEl.addEventListener('click', e => {
-  handleButton(e.target);
-
-  getPopularProductData()
-    .then(array => {
-      console.log(array);
-      renderNewsMarkup(array, valuePage.amountCards);
-    })
-    .catch(error => console.error(error));
-
-  // goToTop();
-});
-
-async function getPopularProductData() {
-  try {
-    const getNews = await getPopularNewsAPI();
-    // console.log('Arr objects with mostpopular News', getNews.results);
-    const newsArr = await getNews.results;
-
-    getAmountCards(newsArr);
-  } catch (error) {
-    notFoundPage.classList.toggle('visually-hidden');
-  }
-  
-  return newsArr;
-}
-
-function getAmountCards(array) {
-  if (window.innerWidth < 768) {
-    valuePage.amountCards = 5;
-    valuePage.totalPages = Math.ceil(array.length / valuePage.amountCards);
-  }
-
-  if (window.innerWidth >= 768 && window.innerWidth < 1280) {
-    valuePage.amountCards = 8;
-    valuePage.totalPages = Math.ceil(array.length / valuePage.amountCards);
-  }
-
-  if (window.innerWidth >= 1280) {
-    valuePage.amountCards = 9;
-    valuePage.totalPages = Math.ceil(array.length / valuePage.amountCards);
-  }
-}
-
-function renderNewsMarkup(data, amountCards) {
-  chunkArray(data, amountCards);
-  // console.log('chunkArray', chunkArray(getNews.results, valuePage.amountCards));
-
-  for (let i = 0; i <= chunkNewsArr.length; i += 1) {
-    console.log(chunkNewsArr[i]);
-    if (valuePage.curPage === i + 1) {
-      console.log('page', chunkNewsArr[i]);
-      createMarkupWithChunkArray(chunkNewsArr[i]);
-      break;
-    }
-  }
-
-  pagination(valuePage);
-
-  handleButtonLeft();
-  handleButtonRight();
-}
-
-function createMarkupWithChunkArray(array) {
-  let markupNews = '';
-  const markupWeather = getMarkupWeather({ data: weatherData });
-  // console.log(markupWeather);
-  // console.log({ data: weatherData });
-  const itemWeather = `<li class="weather__card">${markupWeather.markup}</li>`;
-
- 
-  // console.log(itemWeather);
-
-  if (window.innerWidth < 768) {
-    for (let i = 0; i < array.length; i += 1) {
-      if (i === 0) {
-        markupNews += itemWeather;
-      } else {
-        markupNews += markup(array[i]);
-      }
-    }
-  }
-  if (window.innerWidth >= 768 && window.innerWidth < 1280) {
-    for (let i = 0; i < array.length; i += 1) {
-      if (i === 1) {
-        markupNews += itemWeather;
-      } else {
-        markupNews += markup(array[i]);
-      }
-    }
-  }
-  if (window.innerWidth >= 1280) {
-    for (let i = 0; i < array.length; i += 1) {
-      if (i === 2) {
-        markupNews += itemWeather;
-      } else {
-        markupNews += markup(array[i]);
-      }
-    }
-    if (array.length === 2) {
-      markupNews += itemWeather;
-    }
-  }
-  popularNewsGallery.innerHTML = markupNews;
-  // console.log(markupNews);
-}
+// function markup({ abstract, media, published_date, subsection, title, url }) {
+//   let mediaURL =
+//     'https://amsrus.ru/wp-content/uploads/2016/02/Mercedes-Benz-C63-AMG-Black-Series-1.jpg';
+//   if (
+//     media &&
+//     media[0] &&
+//     media[0]['media-metadata'] &&
+//     media[0]['media-metadata'][2]
+//   ) {
+//     mediaURL = media[0]['media-metadata'][2].url;
+//   }
+//   return `
+//     <li class="card-news__item card__readed">
+//     <div class="card-news__picture"><img src="${mediaURL}" alt="${media[0]?.caption}" class="news-image">
+//       <p class="news-category"> ${subsection}</p>
+//       <button class="news-favorite" aria-label="add to favorite">Add to favorite <svg class="news-favorite__icon" width="16" height="16"><use href="../images/icons-defs.svg#icon-heart-transparent"></use></svg>
+//       </button>
+//     </div>
+//     <div class="card-news__info">
+//       <h3 class="card-news__title"> ${title}</h3>
+//       <p class="card-news__info"> ${abstract}...</p>
+//       <div class="card-information">
+//         <div class="card-infrmation__data"> ${published_date}</div>
+//         <a class="card__infotion__more" href="${url}">Read more</a>
+//       </div>
+//     </div>
+//     <div class="owerlay-readed is-hidden">
+//       <p class="owerlay-readed__info" aria-label="readed">Already read <svg class="owerlay-readed__icon" width="18"  height="18"><use href="../images/icons-defs.svg#icon-readed"></use></svg></p>
+//     </div>
+//   </li>`;
+// }
+// import { getPopularNewsAPI } from './api/news-api.js';
+// // import {newsTemplate} from '../templates/card-news.hbs';
+// import { checkFavCards } from './addAndRemoveFromFavorite.js';​
+// const popularNewsGallery = document.querySelector('.news-gallery');
+// const notFoundPage = document.querySelector('.not-found');​
+// async function getPopularProduct() {
+//   try {
+//     const getNews = await getPopularNewsAPI();
+//     console.log('Arr objects with mostpopular News', getNews.results);
+//     const newsArr = getNews.results;
+//     popularNewsGallery.innerHTML = markup(newsArr);
+//     checkFavCards();
+//   } catch (error) {
+//     notFoundPage.classList.toggle('visually-hidden');
+//   }
+// }​
+// getPopularProduct();
+// function markup(data) {
+//   return data
+//     .map(({ abstract, media, published_date, subsection, title, url }) => {
+//       let mediaURL = `https://webassets.eurac.edu/31538/1647583511-adobestock_490465800.jpeg?auto=format&fm=jpg&h=588&w=980`;
+//       if (
+//         media &&
+//         media[0] &&
+//         media[0]['media-metadata'] &&
+//         media[0]['media-metadata'][2]
+//       ) {
+//         mediaURL = media[0]['media-metadata'][2].url;
+//       }
+//       return `
+//     <li class="card-news__item">
+//     <div class="card-news__picture"><img src="${mediaURL}" alt="${media[0]?.caption}" class="news-image">
+//       <p class="news-category"> ${subsection}</p>
+//       <button class="news-favorite addToFavoriteBtn" aria-label="add to favorite">Add to favorite <svg class="news-favorite__icon" width="16" height="16"><use href="../images/icons-defs.svg#icon-heart-transparent"></use></svg>
+//       </button>
+//     </div>
+//     <div class="card-news__info">
+//       <h3 class="card-news__title"> ${title}</h3>
+//       <p class="card-news__info-chort"> ${abstract}...</p>
+//       <div class="card-information">
+//         <div class="card-infrmation__data"> ${published_date}</div>
+//         <a class="card__infotion__more" rel="nofollow noindex noreferrer" target="_blank" href="${url}">Read more</a>
+//       </div>
+//     </div>
+//     <div class="owerlay-readed">
+//       <p class="owerlay-readed__info" aria-label="readed">Already read <svg class="owerlay-readed__icon" width="18"  height="18"><use href="../images/icons-defs.svg#icon-readed"></use></svg></p>
+//     </div>
+//   </li>`;
+//     })
+//     .join('');
+// }
